Model/requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0

# XAI
lime>=0.2.0.1
scikit-image>=0.22.0

# Web app & API
streamlit>=1.34.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Utilities
Pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.8.0
requests>=2.31.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
opencv-python>=4.9.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4